Add CairoRenderer, move Scene.get_frame()
