+++ tests/helpers/graphical_units.py
@@ -40,7 +40,7 @@ def set_test_scene(scene_object, module_name):
         file_writer_config["text_dir"] = os.path.join(tmpdir, "text")
         file_writer_config["tex_dir"] = os.path.join(tmpdir, "tex")
         scene = scene_object()
-        data = scene.get_frame()
+        data = scene.renderer.get_frame()
 
     tests_directory = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
     path_control_data = os.path.join(
