+++ tests/utils/GraphicalUnitTester.py
@@ -128,7 +128,7 @@ class GraphicalUnitTester:
 
     def test(self, show_diff=False):
         """Compare pre-rendered frame to the frame rendered during the test."""
-        frame_data = self.scene.get_frame()
+        frame_data = self.scene.renderer.get_frame()
         expected_frame_data = self._load_data()
 
         assert frame_data.shape == expected_frame_data.shape, (
